refactor example data to support multiple pathogens.
